delete old remat implementation
moved lowering rule logic from remat_impl.py (now deleted) to ad_checkpoint.py
