+++ jax/_src/lax/control_flow/__init__.py
@@ -19,8 +19,6 @@ from jax._src.lax.control_flow.loops import (associative_scan, cummax, cummax_p,
                                              scan, scan_bind, scan_p,
                                              _scan_impl, while_loop, while_p)
 from jax._src.lax.control_flow.conditionals import cond, cond_p, switch
-from jax._src.lax.control_flow.remat_impl import (remat_impl,
-                                                  optimization_barrier_p)
 from jax._src.lax.control_flow.solves import (custom_linear_solve, custom_root,
                                               _custom_linear_solve_impl,
                                               linear_solve_p)
@@ -32,3 +30,5 @@ from jax._src.lax.control_flow.common import (_initial_style_open_jaxpr,
                                               _initial_style_jaxpr,
                                               _initial_style_jaxprs_with_common_consts,
                                               _check_tree_and_avals)
+# TODO(mattjj): fix dependent library which expects optimization_barrier_p here
+from jax._src.ad_checkpoint import optimization_barrier_p
